Change ambients to a more efficient implementation
Removed Ambient.Provide/Ambient.Consume and ambient() in favor of
Provides() and Ambient<T>.current.

Uses a hash lookup instead of linear scan to determine
the value of an ambient.

Uses a State<T> backed implementation instead to track
reads of ambients instead of custom tracking and invalidation.

Introduces a static ambient that is very cheap to read but very
expensive to change to be used for ambients that rarely or never
change.

Relnote: "Breaking changes to the ambients API. See log and Ambient<T> documentation for details"

BREAKING CHANGES:
  - Ambient keys are now created  with `ambientOf()` or
    `staticAmbientOf()` instead of `Ambient.of()`.
    `Ambient.of()` has been removed.
  - Consume ambients with `Ambient<T>.current` instead of
    using `ambient()`. `ambient()` still works but is
    deprecated.
  - Ambient<T>.Consume was previous discouraged and is now
    removed.
  - Ambients are now provided using `Provides()` instead
    of `Ambient<T>.Provide`. `Ambient<T>.Provide` has been
    removed.